% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbhs}
\alias{gbhs}
\title{Lifebrain Global Brain Health Survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27590 rows and 107 columns.
}
\usage{
data(gbhs)
}
\value{
A data.frame with all responses
}
\description{
Between June 2019 and August 2020, Lifebrain conducted the Global Brain Health Survey to collect data on people’s perceptions of brain health and willingness to take care of their brain by adopting new lifestyles.
The survey was conducted online and translated into 14 languages to reach as many people as possible.
In total, it collected 27,590 responses from people in 81 countries.
}
\examples{
data(gbhs)
gbhs
}
\references{
Budin-Ljøsne I, Friedman BB, Suri S, Solé-Padullés C, Düzel S, Drevon CA, Baaré WFC, Mowinckel AM, Zsoldos E, Madsen KS, Carver RB, Ghisletta P, Arnesen MR, Bartrés Faz D, Brandmaier AM, Fjell AM, Kvalbein A, Henson RN, Kievit RA, Nawijn L, Pochet R, Schnitzler A, Walhovd KB and Zasiekina L (2020) The Global Brain Health Survey: Development of a Multi-Language Survey of Public Views on Brain Health. Front. Public Health 8:387. doi: 10.3389/fpubh.2020.00387
(\href{https://doi.org/10.3389/fpubh.2020.00387}{DOI})
}
\keyword{datasets}
