% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{gbhs_path_child}
\alias{gbhs_path_child}
\title{Get path of child document}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
gbhs_path_child(path = NULL, ...)
}
\arguments{
\item{path}{filename of child document to get path of. If NULL, lists possibilities}

\item{...}{other arguments to \code{\link[base]{list.files}}}
}
\value{
string of file path
}
\description{
Get path of child document
}
\examples{
gbhs_path_child()
gbhs_path_child("bin_desc.Rmd")
gbhs_path_child("ord_mod.Rmd")
}
