% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{gbhs_path_data}
\alias{gbhs_path_data}
\title{Get file path to raw data}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
gbhs_path_data(path = NULL, type = "clean", destination = NULL, ...)
}
\arguments{
\item{path}{Name of file in quotes with extension. If \code{NULL},
the example files will be listed.}

\item{type}{type of data to look up. Either "clean" (default) or
"raw"}

\item{destination}{optional string indicating where to copy the file to}

\item{...}{other arguments to \code{\link[base]{list.files}}}
}
\value{
string of file path
}
\description{
The raw data from the survey is stored in individual files
for each survey language. These are not cleaned or harmonised,
as there are small inconsistencies in coding between the languages.
}
\examples{
gbhs_path_data()
gbhs_path_data("114338_en.tsv")
head(read.delim(gbhs_path_data("114338_en.tsv")))
head(read.delim(gbhs_path_data("114338_en.tsv", "raw")))
}
