% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{gbhs_path_rmd}
\alias{gbhs_path_rmd}
\title{Get file path to template files}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
gbhs_path_rmd(type = "descriptives", paper = 1, destination = NULL)
}
\arguments{
\item{type}{either "descriptive" (default) or "model"}

\item{paper}{an integer of either 1,2 or 3.}

\item{destination}{optional string indicating where to copy the file to}
}
\value{
string of file path
}
\description{
There are two basic type of template files, one descriptive and one with models.
These are based on the exploration and testing of the data towards our publicised
manuscripts and reports. To run the "model" documents, the corresponding "descriptive"
document for that paper must have been previously run.
}
\examples{
gbhs_path_rmd()
gbhs_path_rmd("descriptive", 2)
gbhs_path_rmd("model", 3)
}
