% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{gbhs_path_utilities}
\alias{gbhs_path_utilities}
\title{Get path of utility functions}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
gbhs_path_utilities(path = NULL, ...)
}
\arguments{
\item{path}{filename of utility file to get path of. If NULL, lists possibilities}

\item{...}{other arguments to \code{\link[base]{list.files}}}
}
\value{
string of file path
}
\description{
Get path of utility functions
}
\examples{
gbhs_path_utilities()
gbhs_path_utilities("ggplot-utils.R")
gbhs_path_utilities("data-utils.R")
gbhs_path_utilities("model-utils.R")
}
