% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-reports.R
\name{gbhs_render_report}
\alias{gbhs_render_report}
\title{Render GBHS reports}
\usage{
gbhs_render_report(
  data = gbhs,
  type = "desc",
  paper = 1,
  output_dir = ".",
  ...
)
}
\arguments{
\item{data}{data to be used. Can be a subselection
of the gbhs data, or the entire \code{gbhs} data (default)}

\item{type}{either "descriptive" (default) or "model"}

\item{paper}{an integer of either 1,2 or 3.}

\item{output_dir}{Directory to output the document to.}

\item{...}{other arguments to \code{\link[rmarkdown]{render}}}
}
\value{
creats a report using the data and GBHS template
}
\description{
Descriptives and models for the GBHS data
can be explored by generating the pre-created
report templates.
}
\examples{
\dontrun{
gbhs_render_report(type = "desc", paper = 1)
gbhs_render_report(type = "desc", paper = 2)
gbhs_render_report(type = "mod", paper = 1)
}
}
