% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggmodel}
\alias{ggmodel}
\title{Plot a model output from GBHS}
\usage{
ggmodel(data, y, reverse = FALSE)
}
\arguments{
\item{data}{data as prepared from \code{\link{prep_model_output}}}

\item{y}{What goes on the y-axis}

\item{reverse}{Should the scale be reversed}
}
\value{
ggplot object
}
\description{
Utility function to plot model
output from the GBHS survey
}
