% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggstacked}
\alias{ggstacked}
\title{Create a stacked bar chart}
\usage{
ggstacked(
  data,
  y = key,
  npos = 1.1,
  min_pc = 0.05,
  pattern = NULL,
  n_breaks = 2,
  text_size = 3
)
}
\arguments{
\item{data}{data to plot}

\item{y}{value for the y-axis}

\item{npos}{position of sidebar text}

\item{min_pc}{minimum percent to display text of}

\item{pattern}{regex pattern to use with \code{\link{grepl}} for turning text white/black}

\item{n_breaks}{number of break points}

\item{text_size}{text size}
}
\value{
ggplot object
}
\description{
Create a stacked bar chart
}
