% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggstacked_wrap}
\alias{ggstacked_wrap}
\title{Wrap a ggstacked plot}
\usage{
ggstacked_wrap(data, y, ...)
}
\arguments{
\item{data}{data to wrap}

\item{y}{y-axis variable}

\item{...}{other arguments to \code{\link{ggstacked}}}
}
\value{
ggplot object
}
\description{
Wrap a ggstacked plot
}
