% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit2prob}
\alias{logit2prob}
\title{Transform logit to probability}
\usage{
logit2prob(logit)
}
\arguments{
\item{logit}{a vector of logit scale}
}
\value{
a vector of probabilities
}
\description{
Transform logit to probability
}
\examples{
logit2prob(c(0.5, 1, 1.5))
}
