% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_col_rm}
\alias{na_col_rm}
\title{Remove columns with only NA}
\usage{
na_col_rm(data)
}
\arguments{
\item{data}{data.frame with data}
}
\value{
data.frame without columns that only have NA values
}
\description{
In part of the cleaning process, we needed to
easily remove columns that only contained NA values.
}
\examples{
na_col_rm(mtcars)
}
