% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pc}
\alias{pc}
\title{Calculate percent}
\usage{
pc(x)
}
\arguments{
\item{x}{vector to calculate the percent from}
}
\value{
a vector of percents given the vectors whole.
}
\description{
Calculate percent
}
\examples{
pc(1:10)
}
