% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Pretty percent displaying}
\usage{
percent(x, accuracy = 1, ...)
}
\arguments{
\item{x}{vector of numbers}

\item{accuracy}{accuracy of the percent}

\item{...}{other arguments to \code{\link[scales]{percent}}}
}
\value{
character vector with percentage sign at the end
}
\description{
Pretty percent displaying
}
\examples{
percent(10)
}
