% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{prep_model_output}
\alias{prep_model_output}
\title{Prepare model output for plots}
\usage{
prep_model_output(data, model, y, reverse = FALSE)
}
\arguments{
\item{data}{Data used in the model}

\item{model}{model output}

\item{y}{what goes on the y-axis}

\item{reverse}{whether the categorical scale should be reversed}
}
\value{
fortified data for plotting
}
\description{
Models run come in an output that require a little
work to get plotted. This function helps clean up
and get the data prepared for plotting in particular.
}
