% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surveys}
\alias{surveys}
\title{Survey tibble}
\usage{
surveys()
}
\value{
a tibble with 16 rows and 3 columns
}
\description{
There were in all 16 surveys launched for the
global brain health survey. These were in
different languages to try to capture as many
respondents as possible, especially in Europe.
}
\examples{
surveys()
}
