% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thousand}
\alias{thousand}
\title{Format thousands}
\usage{
thousand(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
character vector where one thousands has a space to the next number
}
\description{
Format thousands
}
\examples{
thousand(c(1000, 150000, 16000))
}
