% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggstacked}
\alias{ggstacked}
\title{Create a stacked bar chart}
\usage{
ggstacked(
  data,
  y = key,
  npos = 1.1,
  min_pc = 0.05,
  n_breaks = 2,
  text_size = 3,
  ...
)
}
\arguments{
\item{data}{data to plot}

\item{y}{value for the y-axis}

\item{npos}{position of sidebar text}

\item{min_pc}{minimum percent to display text of}

\item{n_breaks}{number of break points}

\item{text_size}{text size}

\item{...}{other arguments to mutate}
}
\value{
ggplot object
}
\description{
Create a stacked bar chart
}
